Reorganize context menu item data
